Dialog to add a new ticket from the App
This commit shows how a Widget Class can be extended for your purpose,
how to communicate upstream and add an rpc call to the server.

The TicketDialog extends the normal Dialog by adding a form submission
handler. The syntax to create a Dialog is the same as the one shown
here (i.e. the buttons object, title, etc.). Have a look at the
dialog.js file present in the web addons for more details.

Notice the 'trigger_up' call, which fires a custom Odoo Event up the
widget's hierarchy. This is the preferred way to communicate from child
to parent widget.