Live insertion/deletion from the server
This commit introduces a live notification feature for the creation
deletion of tickets for the connected users.

Note that this functionnality REQUIRES Python code changes.

The controller must now inherit from the BusController class, and
the _poll override should parse the bus polling request to search
for its own parameters then add the relevant channels to the list
of channels listened of that longpolling request. This is of course
mirrored on the JS side, where a specific options is set on the bus
via the line
bus.update_option('demo.ticket', user.partner_id);
Every time a longpolling message must be sent from the Python side,
an explicit bus message must be created using either sendone or
sendmany for the relevant channel.

Note also the synchronicity of events between the main Widget and
the bus: the main Widget sets its listening parameters on the bus
instance, then the Widget starts THEN the bus starts actually polling.

This is important because if the bus starts polling before the
widget has finished starting, a notification might arrive during
the willStart or start of the widget, at which time the dom might
not be ready to be modified by a bus event. It could either cause
an error that would disconnect the bus for up to 50s and/or lose
notification payload silently.

Finally, notice that the bus notification handler is not bound using
events since events necessitate a hierarchy while the bus is completely
otrhogonal to our app widget hierarchy. This is why the handler is
bound in a completely declarative manner.