Add minimal frontend routing
This commit shows an example of how to manage frontend routes,
to allow deep linking your JS app.

It uses a small router library found on a nice JS blog that you may
want to check out:
http://krasimirtsonev.com/blog/

In a nutshell, a Router instance matches a regex with a provided
route and calls a predefined callback when a match is found.